updating active deals
